refactor(expense): modularize expense tracking page components
- Extract expense dialogs into separate components
- Separate table component with its own logic
- Move constants and types to separate files
- Improve code organization and maintainability
