+++ src/lib/expense-constants.ts
@@ -1,0 +1,114 @@
+export const BRANCHES = ['Main Branch', 'Branch 2', 'Branch 3']
+
+export const EXPENSE_TYPES = [
+  { value: 'asset', label: 'Asset' },
+  { value: 'asset-maintenance', label: 'Asset Maintenance' },
+  { value: 'salary', label: 'Salary' },
+  { value: 'utility', label: 'Utility' },
+  { value: 'other', label: 'Other' },
+]
+
+export const SALARY_TYPES = [
+  { value: 'full', label: 'Full' },
+  { value: 'partial', label: 'Partial' },
+  { value: 'advance', label: 'Advance' },
+]
+
+export const PAYMENT_METHODS = [
+  { value: 'cash', label: 'Cash' },
+  { value: 'gcash', label: 'GCash' },
+  { value: 'paymaya', label: 'PayMaya' },
+  { value: 'credit-card', label: 'Credit Card' },
+  { value: 'debit-card', label: 'Debit Card' },
+  { value: 'bank-transfer', label: 'Bank Transfer' },
+  { value: 'online', label: 'Other Online Payment' },
+]
+
+export const SAMPLE_EXPENSES = [
+  {
+    id: '1',
+    type: 'asset',
+    name: 'Treadmill Purchase',
+    date: new Date('2024-01-15'),
+    amount: '45000',
+    branch: 'Main Branch',
+    paymentMethod: 'bank-transfer',
+    category: 'capital',
+    description: 'New cardio equipment',
+    receipt: null,
+  },
+  {
+    id: '2',
+    type: 'utility',
+    name: 'Electricity Bill - January',
+    date: new Date('2024-01-10'),
+    amount: '8500',
+    branch: 'Main Branch',
+    paymentMethod: 'bank-transfer',
+    category: 'operational',
+    description: 'Monthly electricity payment',
+    receipt: null,
+  },
+  {
+    id: '3',
+    type: 'salary',
+    name: 'Trainer Salary - January',
+    date: new Date('2024-01-05'),
+    amount: '25000',
+    branch: 'Main Branch',
+    paymentMethod: 'bank-transfer',
+    category: 'operational',
+    description: 'Monthly salary payment',
+    receipt: null,
+    salaryType: 'full',
+  },
+  {
+    id: '4',
+    type: 'asset-maintenance',
+    name: 'Equipment Repair',
+    date: new Date('2024-01-20'),
+    amount: '3500',
+    branch: 'Branch 2',
+    paymentMethod: 'cash',
+    category: 'operational',
+    description: 'Repair of weight machines',
+    receipt: null,
+  },
+  {
+    id: '5',
+    type: 'other',
+    name: 'Cleaning Supplies',
+    date: new Date('2024-01-12'),
+    amount: '1200',
+    branch: 'Main Branch',
+    paymentMethod: 'cash',
+    category: 'operational',
+    description: 'Monthly cleaning supplies stock',
+    receipt: null,
+  },
+  {
+    id: '6',
+    type: 'salary',
+    name: 'Staff Advance Payment',
+    date: new Date('2023-12-28'),
+    amount: '5000',
+    branch: 'Main Branch',
+    paymentMethod: 'gcash',
+    category: 'operational',
+    description: 'Advance payment for staff',
+    receipt: null,
+    salaryType: 'advance',
+  },
+  {
+    id: '7',
+    type: 'utility',
+    name: 'Water Bill - December',
+    date: new Date('2023-12-15'),
+    amount: '2800',
+    branch: 'Main Branch',
+    paymentMethod: 'bank-transfer',
+    category: 'operational',
+    description: 'Monthly water payment',
+    receipt: null,
+  },
+]
