+++ src/lib/expense-types.ts
@@ -1,0 +1,13 @@
+export interface ExpenseFormData {
+  id: string
+  type: string
+  name: string
+  date: Date
+  amount: string
+  branch: string
+  paymentMethod: string
+  category: string
+  description: string
+  receipt: File | null
+  salaryType?: string
+}
